Organized the scripts/profile folder and removed non-route functions from profileRoutes.
