+++ public/scripts/profile/account.js
@@ -1,33 +1,32 @@
 
 
-
 // Improved function to add Edit Availability button
 function addAvailabilityButton() {
-  // Find all section titles
-  const sectionTitles = document.querySelectorAll('.section-title');
+    // Find all section titles
+    const sectionTitles = document.querySelectorAll('.section-title');
 
-  sectionTitles.forEach(title => {
-    if (title.textContent.includes('Weekly Availability')) {
-      const button = document.createElement('a');
-      button.href = '/set-availability';
-      button.className = 'btn btn-availability ms-3';
-      button.innerHTML = '<i class="bi bi-pencil"></i> Edit';
+    sectionTitles.forEach(title => {
+        if (title.textContent.includes('Weekly Availability')) {
+            const button = document.createElement('a');
+            button.href = '/set-availability';
+            button.className = 'btn btn-availability ms-3';
+            button.innerHTML = '<i class="bi bi-pencil"></i> Edit';
 
-      // Create wrapper for proper alignment
-      const wrapper = document.createElement('div');
-      wrapper.className = 'd-flex align-items-center justify-content-between mb-3';
+            // Create wrapper for proper alignment
+            const wrapper = document.createElement('div');
+            wrapper.className = 'd-flex align-items-center justify-content-between mb-3';
 
-      // Reorganize DOM elements
-      const parent = title.parentNode;
-      parent.insertBefore(wrapper, title);
-      wrapper.appendChild(title);
-      wrapper.appendChild(button);
-    }
-  });
+            // Reorganize DOM elements
+            const parent = title.parentNode;
+            parent.insertBefore(wrapper, title);
+            wrapper.appendChild(title);
+            wrapper.appendChild(button);
+
+        }
+    });
 }
 
 function getUserLocation() {
-  console.log("Getting user location...");
 
   if ("geolocation" in navigator) {
     navigator.geolocation.getCurrentPosition(
@@ -66,6 +65,7 @@ function getUserLocation() {
 
 // Initialize everything when DOM loads
 document.addEventListener('DOMContentLoaded', function () {
+  console.log('DOM Content Loaded');
   addAvailabilityButton();
   getUserLocation();
 });
