+++ public/scripts/profile/scheduleController.js
