+++ views/account.ejs
@@ -1,34 +1,44 @@
 <%- include('templates/header',{css:["css/account.css"]}) %>
 
 <!-- Include Bootstrap Icons -->
-<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.5/font/bootstrap-icons.css">
+<link
+  rel="stylesheet"
+  href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.5/font/bootstrap-icons.css"
+/>
 
 <div class="container py-4">
-    <!-- Action Buttons Row -->
-    <div class="d-flex justify-content-center gap-3 mb-4 flex-wrap">
-        <!-- Meetings Button -->
-        <a href="/connect/meetings" class="btn btn-action d-flex align-items-center gap-2">
-            <i class="bi bi-calendar-event"></i>
-            Meetings
-        </a>
+  <!-- Action Buttons Row -->
+  <div class="d-flex justify-content-center gap-3 mb-4 flex-wrap">
+    <!-- Meetings Button -->
+    <a
+      href="/connect/meetings"
+      class="btn btn-action d-flex align-items-center gap-2"
+    >
+      <i class="bi bi-calendar-event"></i>
+      Meetings
+    </a>
 
-        <!-- Edit Profile Button -->
-        <a href="/create-profile" class="btn btn-action d-flex align-items-center gap-2">
-            <i class="bi bi-pencil"></i>
-            Edit Profile
-        </a>
+    <!-- Edit Profile Button -->
+    <a
+      href="/create-profile"
+      class="btn btn-action d-flex align-items-center gap-2"
+    >
+      <i class="bi bi-pencil"></i>
+      Edit Profile
+    </a>
 
-        <!-- Logout Button -->
-        <a href="/logout" class="btn btn-logout d-flex align-items-center gap-2">
-            <i class="bi bi-box-arrow-right"></i>
-            Logout
-        </a>
-    </div>
+    <!-- Logout Button -->
+    <a href="/logout" class="btn btn-logout d-flex align-items-center gap-2">
+      <i class="bi bi-box-arrow-right"></i>
+      Logout
+    </a>
+  </div>
 
-    <!-- Profile Section -->
-    <%- include('templates/profile', {userData:userData}) %>
+  <!-- Profile Section -->
+  <%- include('templates/profile', {userData:userData}) %>
 </div>
-<script src="/scripts/account.js"></script>
 
+<!-- Load scripts at the end of body -->
+<script src="/scripts/profile/account.js" defer></script>
 
 <%- include('templates/footer', {currentPage:'account'}) %>
