+++ views/createProfile.ejs
@@ -89,7 +89,7 @@
   </div>
 
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
-  <script src="/scripts/fieldsetController.js"></script>
+  <script src="/scripts/profile/fieldsetController.js"></script>
 
 
   <script>
